feat: move get_adb_id
