+++ modelconverter/utils/adb_handler.py
@@ -3,7 +3,6 @@ import subprocess
 
 from loguru import logger
 from luxonis_ml.typing import PathType
-import depthai as dai
 
 
 class AdbHandler:
@@ -78,14 +77,3 @@ class AdbHandler:
         logger.info(f"Using device ID: {device_id}")
 
         return device_id
-
-
-def get_adb_id(device_ip: str) -> str:
-    if device_ip is None:
-        return None
-    try:
-        with dai.Device(device_ip) as device:
-            device_mxid = int(device.getDeviceId())
-    except:
-        raise RuntimeError(f"Failed to connect to device: {device_ip}")
-    return format(device_mxid, "x")  # abd_id is hex version of mxid
